Shift timer logic to App and push the user to start screen after timer is expired
